BUNK3R_IA/requirements.txt
# BUNK3R_IA - Dependencias del sistema de IA
# Versión: 1.0.0

# Core Flask
flask>=2.3.0
gunicorn>=21.0.0

# Database
psycopg2-binary>=2.9.0

# AI Providers
requests>=2.31.0
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0

# Optional AI Provider SDKs (install as needed)
# groq>=0.4.0
# google-generativeai>=0.3.0
# huggingface-hub>=0.19.0
